feat(dashboard): add order management
